Login and Session

+++ src/main/java/com/simplilearn/fitnessclubautomation/controller/SubscriberController.java
@@ -14,6 +14,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -68,7 +70,7 @@ public class SubscriberController {
     }
 
     @RequestMapping(value = "/subscriber/edit", method = RequestMethod.POST)
-    private String editSubscriber(@RequestParam(value = "subscriber_id", required = true) Long subscriber_id,
+    private String editSubscriber(ModelMap modelMap, HttpServletRequest request, @RequestParam(value = "subscriber_id", required = true) Long subscriber_id,
                                   @RequestParam(value = "subscriber_name", required = true) String subscriber_name,
                                   @RequestParam(value = "subscriber_age", required = true) int subscriber_age,
                                   @RequestParam(value = "subscriber_gender", required = true) String subscriber_gender,
@@ -76,6 +78,20 @@ public class SubscriberController {
                                   @RequestParam(value = "subscriber_address", required = true) String subscriber_address,
                                   @RequestParam(value = "subscriber_trainer_id", required = true) Long subscriber_trainer_id,
                                   @RequestParam(value = "subscription_plan_id", required = true) Long subscription_plan_id) {
+        HttpSession session = request.getSession();
+        if (session.getAttribute("adminId") == null) {
+            return "redirect:/";
+        }
+        if (subscriber_trainer_id == 0) {
+            modelMap.addAttribute("error", true);
+            modelMap.addAttribute("message", "Please select a valid trainer");
+            return "edit-subscriber";
+        }
+        if (subscription_plan_id == 0) {
+            modelMap.addAttribute("error", true);
+            modelMap.addAttribute("message", "Please select a valid subscription plan");
+            return "edit-subscriber";
+        }
         Trainer trainer = trainerService.getTrainer(subscriber_trainer_id);
         SubscriptionPlan subscriptionPlan = subscriptionPlanService.getSubscriptionPlan(subscription_plan_id);
         int paid_fees = subscriberService.getSubscriber(subscriber_id).getSubscriberFeesPaid();
@@ -85,33 +101,45 @@ public class SubscriberController {
     }
 
     @RequestMapping(value = "/subscriber/delete/{subscriber_id}")
-    private Boolean deleteSubscriber(@PathVariable Long subscriber_id) {
+    private String deleteSubscriber(@PathVariable Long subscriber_id, HttpServletRequest request) {
+        HttpSession session = request.getSession();
+        if (session.getAttribute("adminId") == null) {
+            return "redirect:/";
+        }
         try {
             subscriberService.deleteSubscriber(subscriber_id);
-            return true;
+            return "redirect:/subscriber";
         } catch (Exception ex) {
-            return false;
+            return "redirect:/subscriber";
         }
     }
 
     @RequestMapping(value = "/subscriber")
-    private String getAllSubscribers(ModelMap modelMap) {
+    private String getAllSubscribers(ModelMap modelMap, HttpServletRequest request) {
+        HttpSession session = request.getSession();
+        if (session.getAttribute("adminId") == null) {
+            return "redirect:/";
+        }
         List<Subscriber> subscribers = new ArrayList<>();
         try {
             subscribers = subscriberService.getAllSubscribers();
             modelMap.addAttribute("subscriber_list", subscribers);
             modelMap.addAttribute("message", "Total <b>" + subscribers.size() + "</b> subscribers found.");
             return "subscriber-list";
-//            return subscribers;
         } catch (Exception ex) {
-//            return subscribers;
+            modelMap.addAttribute("error", true);
+            modelMap.addAttribute("message", ex.getMessage() + "");
             return "subscriber-list";
         }
     }
 
     @RequestMapping(value = "/subscriber/{subscriber_id}")
-    private String getSubscriber(ModelMap modelMap, @PathVariable Long subscriber_id) {
+    private String getSubscriber(ModelMap modelMap, HttpServletRequest request, @PathVariable Long subscriber_id) {
         Subscriber subscriber = new Subscriber();
+        HttpSession session = request.getSession();
+        if (session.getAttribute("adminId") == null) {
+            return "redirect:/";
+        }
         try {
             subscriber = subscriberService.getSubscriber(subscriber_id);
             modelMap.addAttribute("subscriber", subscriber);
